pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shodan-tui"
version = "0.1.0"
description = "A terminal user interface for Shodan — built for OSINT workflows"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "textual>=0.50.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
shodan-tui = "main:main"

[project.optional-dependencies]
dev = [
    "textual-dev>=0.50.0",
]

[tool.hatch.build.targets.wheel]
packages = ["shodan_tui"]
